% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributeCapabilityPlan.R
\name{AttributeCapabilityPlotter}
\alias{AttributeCapabilityPlotter}
\title{AttributeCapabilityPlotter}
\description{
Thin delegation plotter for attribute capability analysis
(binomial / poisson). Dispatches every panel to the matching
\verb{iQualityR.plot::plot_attribute_*} function. No ggplot2 geom_* / stat_*
logic lives here.
}
\details{
AttributeCapabilityPlotter
}
\section{Super class}{
\code{\link[iQualityR.core:IqrPlotterBase]{iQualityR.core::IqrPlotterBase}} -> \code{AttributeCapabilityPlotter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttributeCapabilityPlotter-render}{\code{AttributeCapabilityPlotter$render()}}
\item \href{#method-AttributeCapabilityPlotter-.plot_sixpack}{\code{AttributeCapabilityPlotter$.plot_sixpack()}}
\item \href{#method-AttributeCapabilityPlotter-.plot_control}{\code{AttributeCapabilityPlotter$.plot_control()}}
\item \href{#method-AttributeCapabilityPlotter-.plot_rate}{\code{AttributeCapabilityPlotter$.plot_rate()}}
\item \href{#method-AttributeCapabilityPlotter-.plot_cumulative}{\code{AttributeCapabilityPlotter$.plot_cumulative()}}
\item \href{#method-AttributeCapabilityPlotter-.plot_defects}{\code{AttributeCapabilityPlotter$.plot_defects()}}
\item \href{#method-AttributeCapabilityPlotter-.plot_fit}{\code{AttributeCapabilityPlotter$.plot_fit()}}
\item \href{#method-AttributeCapabilityPlotter-.plot_performance}{\code{AttributeCapabilityPlotter$.plot_performance()}}
\item \href{#method-AttributeCapabilityPlotter-.effective_target}{\code{AttributeCapabilityPlotter$.effective_target()}}
\item \href{#method-AttributeCapabilityPlotter-.build_subtitle}{\code{AttributeCapabilityPlotter$.build_subtitle()}}
\item \href{#method-AttributeCapabilityPlotter-clone}{\code{AttributeCapabilityPlotter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".contrast_text"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.contrast_text'><code>iQualityR.core::IqrPlotterBase$.contrast_text()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_discrete"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_discrete'><code>iQualityR.core::IqrPlotterBase$.pal_discrete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_diverging"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_diverging'><code>iQualityR.core::IqrPlotterBase$.pal_diverging()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_semantic"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_semantic'><code>iQualityR.core::IqrPlotterBase$.pal_semantic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_sequential"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_sequential'><code>iQualityR.core::IqrPlotterBase$.pal_sequential()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_ui"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_ui'><code>iQualityR.core::IqrPlotterBase$.pal_ui()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_discrete"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_discrete'><code>iQualityR.core::IqrPlotterBase$.scale_color_discrete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_diverging"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_diverging'><code>iQualityR.core::IqrPlotterBase$.scale_color_diverging()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_semantic"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_semantic'><code>iQualityR.core::IqrPlotterBase$.scale_color_semantic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_sequential"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_sequential'><code>iQualityR.core::IqrPlotterBase$.scale_color_sequential()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_color_paired"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_color_paired'><code>iQualityR.core::IqrPlotterBase$.scale_fill_color_paired()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_discrete"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_discrete'><code>iQualityR.core::IqrPlotterBase$.scale_fill_discrete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_diverging"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_diverging'><code>iQualityR.core::IqrPlotterBase$.scale_fill_diverging()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_semantic"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_semantic'><code>iQualityR.core::IqrPlotterBase$.scale_fill_semantic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_sequential"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_sequential'><code>iQualityR.core::IqrPlotterBase$.scale_fill_sequential()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-render"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-render}{}}}
\subsection{Method \code{render()}}{
Render a plot by type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$render(results, theme_obj, type = "full", plan, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Analysis results list.}

\item{\code{theme_obj}}{IqrTheme object.}

\item{\code{type}}{Plot type: "full" (Sixpack), "control", "histogram"
(alias for "rate"), "rate", "cumulative", "defects", "fit",
"performance".}

\item{\code{plan}}{AttributeCapabilityPlan object.}

\item{\code{...}}{Additional arguments (unused).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.plot_sixpack"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.plot_sixpack}{}}}
\subsection{Method \code{.plot_sixpack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.plot_sixpack(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.plot_control"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.plot_control}{}}}
\subsection{Method \code{.plot_control()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.plot_control(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.plot_rate"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.plot_rate}{}}}
\subsection{Method \code{.plot_rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.plot_rate(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.plot_cumulative"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.plot_cumulative}{}}}
\subsection{Method \code{.plot_cumulative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.plot_cumulative(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.plot_defects"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.plot_defects}{}}}
\subsection{Method \code{.plot_defects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.plot_defects(results, theme_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.plot_fit"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.plot_fit}{}}}
\subsection{Method \code{.plot_fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.plot_fit(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.plot_performance"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.plot_performance}{}}}
\subsection{Method \code{.plot_performance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.plot_performance(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.effective_target"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.effective_target}{}}}
\subsection{Method \code{.effective_target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.effective_target(results, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-.build_subtitle"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-.build_subtitle}{}}}
\subsection{Method \code{.build_subtitle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$.build_subtitle(results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeCapabilityPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeCapabilityPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeCapabilityPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
