% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributeCapabilityPlan.R
\name{IqrAttributeCapabilityTask}
\alias{IqrAttributeCapabilityTask}
\title{IqrAttributeCapabilityTask}
\description{
Task coordinator for attribute capability analysis. Inherits
\code{IqrTaskBase} and dispatches to \code{AttributeCapabilityAnalyzer} /
\code{AttributeCapabilityPlotter}.
}
\details{
IqrAttributeCapabilityTask
}
\section{Super class}{
\code{\link[iQualityR.core:IqrTaskBase]{iQualityR.core::IqrTaskBase}} -> \code{IqrAttributeCapabilityTask}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IqrAttributeCapabilityTask-new}{\code{IqrAttributeCapabilityTask$new()}}
\item \href{#method-IqrAttributeCapabilityTask-compute}{\code{IqrAttributeCapabilityTask$compute()}}
\item \href{#method-IqrAttributeCapabilityTask-summary}{\code{IqrAttributeCapabilityTask$summary()}}
\item \href{#method-IqrAttributeCapabilityTask-plot}{\code{IqrAttributeCapabilityTask$plot()}}
\item \href{#method-IqrAttributeCapabilityTask-report}{\code{IqrAttributeCapabilityTask$report()}}
\item \href{#method-IqrAttributeCapabilityTask-clone}{\code{IqrAttributeCapabilityTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAttributeCapabilityTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAttributeCapabilityTask-new}{}}}
\subsection{Method \code{new()}}{
Create a task instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAttributeCapabilityTask$new(
  data,
  defects,
  sample_sizes,
  plan,
  theme = "academic",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing defect counts and sample sizes.}

\item{\code{defects}}{Column name for defect/defective counts.}

\item{\code{sample_sizes}}{Column name for subgroup sizes / exposures.}

\item{\code{plan}}{\link{AttributeCapabilityPlan} object.}

\item{\code{theme}}{Theme name or IqrTheme object.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAttributeCapabilityTask-compute"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAttributeCapabilityTask-compute}{}}}
\subsection{Method \code{compute()}}{
Execute attribute capability analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAttributeCapabilityTask$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAttributeCapabilityTask-summary"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAttributeCapabilityTask-summary}{}}}
\subsection{Method \code{summary()}}{
Print summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAttributeCapabilityTask$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAttributeCapabilityTask-plot"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAttributeCapabilityTask-plot}{}}}
\subsection{Method \code{plot()}}{
Render plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAttributeCapabilityTask$plot(type = "full", theme = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{Theme name or IqrTheme object.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAttributeCapabilityTask-report"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAttributeCapabilityTask-report}{}}}
\subsection{Method \code{report()}}{
Generate Excel report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAttributeCapabilityTask$report(format = "excel", path = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrAttributeCapabilityTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IqrAttributeCapabilityTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrAttributeCapabilityTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
