% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivarCapabilityPlan.R
\name{IqrMultivarCapabilityTask}
\alias{IqrMultivarCapabilityTask}
\title{IqrMultivarCapabilityTask}
\description{
Task coordinator for multivariate capability analysis.
Inherits \code{IqrTaskBase}.
}
\details{
IqrMultivarCapabilityTask
}
\section{Super class}{
\code{\link[iQualityR.core:IqrTaskBase]{iQualityR.core::IqrTaskBase}} -> \code{IqrMultivarCapabilityTask}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IqrMultivarCapabilityTask-new}{\code{IqrMultivarCapabilityTask$new()}}
\item \href{#method-IqrMultivarCapabilityTask-compute}{\code{IqrMultivarCapabilityTask$compute()}}
\item \href{#method-IqrMultivarCapabilityTask-summary}{\code{IqrMultivarCapabilityTask$summary()}}
\item \href{#method-IqrMultivarCapabilityTask-plot}{\code{IqrMultivarCapabilityTask$plot()}}
\item \href{#method-IqrMultivarCapabilityTask-report}{\code{IqrMultivarCapabilityTask$report()}}
\item \href{#method-IqrMultivarCapabilityTask-clone}{\code{IqrMultivarCapabilityTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrMultivarCapabilityTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-IqrMultivarCapabilityTask-new}{}}}
\subsection{Method \code{new()}}{
Create a task instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrMultivarCapabilityTask$new(data, ctqs, plan, theme = "academic", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame with CTQ columns.}

\item{\code{ctqs}}{Character vector of CTQ column names.}

\item{\code{plan}}{\link{MultivarCapabilityPlan} object.}

\item{\code{theme}}{Theme name or IqrTheme object.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrMultivarCapabilityTask-compute"></a>}}
\if{latex}{\out{\hypertarget{method-IqrMultivarCapabilityTask-compute}{}}}
\subsection{Method \code{compute()}}{
Execute multivariate capability analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrMultivarCapabilityTask$compute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrMultivarCapabilityTask-summary"></a>}}
\if{latex}{\out{\hypertarget{method-IqrMultivarCapabilityTask-summary}{}}}
\subsection{Method \code{summary()}}{
Print summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrMultivarCapabilityTask$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrMultivarCapabilityTask-plot"></a>}}
\if{latex}{\out{\hypertarget{method-IqrMultivarCapabilityTask-plot}{}}}
\subsection{Method \code{plot()}}{
Render plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrMultivarCapabilityTask$plot(type = "full", theme = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{Theme name or IqrTheme object.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrMultivarCapabilityTask-report"></a>}}
\if{latex}{\out{\hypertarget{method-IqrMultivarCapabilityTask-report}{}}}
\subsection{Method \code{report()}}{
Generate Excel report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrMultivarCapabilityTask$report(format = "excel", path = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IqrMultivarCapabilityTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IqrMultivarCapabilityTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IqrMultivarCapabilityTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
