% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{iQualityR.capa-package}
\alias{iQualityR.capa}
\alias{iQualityR.capa-package}
\title{iQualityR.capa}
\description{
Process capability analysis tools for the iQualityR ecosystem.
}
\author{
\strong{Maintainer}: iQualityR Team \email{iQualityR@example.com}

}
\keyword{internal}
