% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivarCapabilityPlan.R
\name{MultivarCapabilityPlan}
\alias{MultivarCapabilityPlan}
\title{MultivarCapabilityPlan}
\description{
Configuration for multivariate process capability analysis.
Inherits \code{IqrPlanBase}.
}
\details{
MultivarCapabilityPlan
}
\section{Super class}{
\code{\link[iQualityR.core:IqrPlanBase]{iQualityR.core::IqrPlanBase}} -> \code{MultivarCapabilityPlan}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lsl_vec}}{Numeric vector of lower specs (length p).}

\item{\code{usl_vec}}{Numeric vector of upper specs (length p).}

\item{\code{target_vec}}{Optional target vector.}

\item{\code{method}}{Reserved for future: \code{"both"} (default), \code{"mcpv"}, \code{"hpci"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultivarCapabilityPlan-new}{\code{MultivarCapabilityPlan$new()}}
\item \href{#method-MultivarCapabilityPlan-to_list}{\code{MultivarCapabilityPlan$to_list()}}
\item \href{#method-MultivarCapabilityPlan-clone}{\code{MultivarCapabilityPlan$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="generate_protocol"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-generate_protocol'><code>iQualityR.core::IqrPlanBase$generate_protocol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="set_criteria"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-set_criteria'><code>iQualityR.core::IqrPlanBase$set_criteria()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="set_meta"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-set_meta'><code>iQualityR.core::IqrPlanBase$set_meta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlanBase" data-id="validate"><a href='../../iQualityR.core/html/IqrPlanBase.html#method-IqrPlanBase-validate'><code>iQualityR.core::IqrPlanBase$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivarCapabilityPlan-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultivarCapabilityPlan-new}{}}}
\subsection{Method \code{new()}}{
Create a new multivariate capability plan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivarCapabilityPlan$new(
  lsl_vec,
  usl_vec,
  target_vec = NULL,
  method = "both",
  conf_level = 0.95,
  task_tag = "multivar_capability",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lsl_vec}}{Numeric vector of lower specs.}

\item{\code{usl_vec}}{Numeric vector of upper specs.}

\item{\code{target_vec}}{Optional target vector.}

\item{\code{method}}{Reserved: \code{"both"}, \code{"mcpv"}, or \code{"hpci"}.}

\item{\code{conf_level}}{Confidence level.}

\item{\code{task_tag}}{Task tag.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivarCapabilityPlan-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-MultivarCapabilityPlan-to_list}{}}}
\subsection{Method \code{to_list()}}{
Export configuration as a list (overrides base method)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivarCapabilityPlan$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivarCapabilityPlan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultivarCapabilityPlan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivarCapabilityPlan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
