% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivarCapabilityPlan.R
\name{MultivariateCapabilityPlotter}
\alias{MultivariateCapabilityPlotter}
\title{MultivariateCapabilityPlotter}
\description{
Thin delegation plotter for multivariate process capability
analysis (MCPV / HPCI). Dispatches every panel to the matching
\verb{iQualityR.plot::plot_*} function. No ggplot2 geom_* / stat_* logic
lives here.
}
\details{
MultivariateCapabilityPlotter
}
\section{Super class}{
\code{\link[iQualityR.core:IqrPlotterBase]{iQualityR.core::IqrPlotterBase}} -> \code{MultivariateCapabilityPlotter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultivariateCapabilityPlotter-render}{\code{MultivariateCapabilityPlotter$render()}}
\item \href{#method-MultivariateCapabilityPlotter-.plot_sixpack}{\code{MultivariateCapabilityPlotter$.plot_sixpack()}}
\item \href{#method-MultivariateCapabilityPlotter-.plot_ellipse}{\code{MultivariateCapabilityPlotter$.plot_ellipse()}}
\item \href{#method-MultivariateCapabilityPlotter-.plot_mcpv_bar}{\code{MultivariateCapabilityPlotter$.plot_mcpv_bar()}}
\item \href{#method-MultivariateCapabilityPlotter-.plot_marginal}{\code{MultivariateCapabilityPlotter$.plot_marginal()}}
\item \href{#method-MultivariateCapabilityPlotter-.plot_t2}{\code{MultivariateCapabilityPlotter$.plot_t2()}}
\item \href{#method-MultivariateCapabilityPlotter-.plot_mcpv}{\code{MultivariateCapabilityPlotter$.plot_mcpv()}}
\item \href{#method-MultivariateCapabilityPlotter-.plot_joint_ppm}{\code{MultivariateCapabilityPlotter$.plot_joint_ppm()}}
\item \href{#method-MultivariateCapabilityPlotter-.build_subtitle}{\code{MultivariateCapabilityPlotter$.build_subtitle()}}
\item \href{#method-MultivariateCapabilityPlotter-.placeholder}{\code{MultivariateCapabilityPlotter$.placeholder()}}
\item \href{#method-MultivariateCapabilityPlotter-clone}{\code{MultivariateCapabilityPlotter$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".contrast_text"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.contrast_text'><code>iQualityR.core::IqrPlotterBase$.contrast_text()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_discrete"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_discrete'><code>iQualityR.core::IqrPlotterBase$.pal_discrete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_diverging"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_diverging'><code>iQualityR.core::IqrPlotterBase$.pal_diverging()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_semantic"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_semantic'><code>iQualityR.core::IqrPlotterBase$.pal_semantic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_sequential"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_sequential'><code>iQualityR.core::IqrPlotterBase$.pal_sequential()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".pal_ui"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.pal_ui'><code>iQualityR.core::IqrPlotterBase$.pal_ui()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_discrete"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_discrete'><code>iQualityR.core::IqrPlotterBase$.scale_color_discrete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_diverging"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_diverging'><code>iQualityR.core::IqrPlotterBase$.scale_color_diverging()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_semantic"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_semantic'><code>iQualityR.core::IqrPlotterBase$.scale_color_semantic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_color_sequential"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_color_sequential'><code>iQualityR.core::IqrPlotterBase$.scale_color_sequential()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_color_paired"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_color_paired'><code>iQualityR.core::IqrPlotterBase$.scale_fill_color_paired()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_discrete"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_discrete'><code>iQualityR.core::IqrPlotterBase$.scale_fill_discrete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_diverging"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_diverging'><code>iQualityR.core::IqrPlotterBase$.scale_fill_diverging()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_semantic"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_semantic'><code>iQualityR.core::IqrPlotterBase$.scale_fill_semantic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrPlotterBase" data-id=".scale_fill_sequential"><a href='../../iQualityR.core/html/IqrPlotterBase.html#method-IqrPlotterBase-.scale_fill_sequential'><code>iQualityR.core::IqrPlotterBase$.scale_fill_sequential()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-render"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-render}{}}}
\subsection{Method \code{render()}}{
Render a plot by type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$render(
  results,
  theme_obj,
  type = "full",
  plan,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{Analysis results list.}

\item{\code{theme_obj}}{IqrTheme object.}

\item{\code{type}}{Plot type: "full" (Sixpack), "ellipse", "mcpv_bar",
"marginal", "t2", "mcpv", "joint_ppm".}

\item{\code{plan}}{MultivarCapabilityPlan object.}

\item{\code{...}}{Additional arguments (unused).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.plot_sixpack"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.plot_sixpack}{}}}
\subsection{Method \code{.plot_sixpack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.plot_sixpack(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.plot_ellipse"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.plot_ellipse}{}}}
\subsection{Method \code{.plot_ellipse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.plot_ellipse(results, theme_obj, plan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.plot_mcpv_bar"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.plot_mcpv_bar}{}}}
\subsection{Method \code{.plot_mcpv_bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.plot_mcpv_bar(results, theme_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.plot_marginal"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.plot_marginal}{}}}
\subsection{Method \code{.plot_marginal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.plot_marginal(results, theme_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.plot_t2"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.plot_t2}{}}}
\subsection{Method \code{.plot_t2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.plot_t2(results, theme_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.plot_mcpv"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.plot_mcpv}{}}}
\subsection{Method \code{.plot_mcpv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.plot_mcpv(results, theme_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.plot_joint_ppm"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.plot_joint_ppm}{}}}
\subsection{Method \code{.plot_joint_ppm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.plot_joint_ppm(results, theme_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.build_subtitle"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.build_subtitle}{}}}
\subsection{Method \code{.build_subtitle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.build_subtitle(results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-.placeholder"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-.placeholder}{}}}
\subsection{Method \code{.placeholder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$.placeholder(results, theme_obj, title, message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultivariateCapabilityPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MultivariateCapabilityPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultivariateCapabilityPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
